+++ mods/GUIs/modsUI/mods.h
@@ -6,8 +6,8 @@
 //
 
 #define MODS_SESSION_NAME "MODS Control Panel" //!< Window Name for banner
-#define MODS_REV_NUMBER   "v3.2.0-archon"      //!< MODS UI Revision number
-#define MODS_REV_DATE     "2026 Jan 18"        //!< Revision date
+#define MODS_REV_NUMBER   "v3.2.1-archon"      //!< MODS UI Revision number
+#define MODS_REV_DATE     "2026 Jan 19"        //!< Revision date
 
 // Runtime files and paths